+++ common/cosmwasm-smart-contracts/mixnet-contract/src/reward_params.rs
@@ -27,6 +27,11 @@ pub struct IntervalRewardParams {
     #[cfg_attr(feature = "generate-ts", ts(type = "string"))]
     pub staking_supply: Decimal,
 
+    /// Defines the percentage of stake needed to reach saturation for all of the nodes in the rewarded set.
+    /// Also known as `beta`.
+    #[cfg_attr(feature = "generate-ts", ts(type = "string"))]
+    pub staking_supply_scale_factor: Percent,
+
     // computed values
     /// Current value of the computed reward budget per epoch, per node.
     /// It is expected to be constant throughout the interval.
@@ -174,6 +179,10 @@ impl RewardingParams {
             self.interval.staking_supply = staking_supply;
         }
 
+        if let Some(staking_supply_scale_factor) = updates.staking_supply_scale_factor {
+            self.interval.staking_supply_scale_factor = staking_supply_scale_factor
+        }
+
         if let Some(sybil_resistance_percent) = updates.sybil_resistance_percent {
             self.interval.sybil_resistance = sybil_resistance_percent;
         }
@@ -246,6 +255,9 @@ pub struct IntervalRewardingParamsUpdate {
     pub staking_supply: Option<Decimal>,
 
     #[cfg_attr(feature = "generate-ts", ts(type = "string | null"))]
+    pub staking_supply_scale_factor: Option<Percent>,
+
+    #[cfg_attr(feature = "generate-ts", ts(type = "string | null"))]
     pub sybil_resistance_percent: Option<Percent>,
 
     #[cfg_attr(feature = "generate-ts", ts(type = "string | null"))]
@@ -262,6 +274,7 @@ impl IntervalRewardingParamsUpdate {
         // essentially at least a single field has to be a `Some`
         self.reward_pool.is_some()
             || self.staking_supply.is_some()
+            || self.staking_supply_scale_factor.is_some()
             || self.sybil_resistance_percent.is_some()
             || self.active_set_work_factor.is_some()
             || self.interval_pool_emission.is_some()