+++ common/cosmwasm-smart-contracts/vesting-contract/src/lib.rs
@@ -4,10 +4,13 @@
 #![warn(clippy::expect_used)]
 #![warn(clippy::unwrap_used)]
 
+use contracts_common::Percent;
 use cosmwasm_std::{Addr, Coin, Timestamp, Uint128};
+use log::warn;
 use mixnet_contract_common::MixId;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+use std::str::FromStr;
 
 pub use messages::{ExecuteMsg, InitMsg, MigrateMsg, QueryMsg};
 
@@ -43,6 +46,47 @@ impl PledgeData {
 
     pub fn new(amount: Coin, block_time: Timestamp) -> Self {
         Self { amount, block_time }
+    }
+}
+
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub enum PledgeCap {
+    Percent(Percent),
+    Absolute(Uint128), // This has to be in unym
+}
+
+impl FromStr for PledgeCap {
+    type Err = String;
+
+    fn from_str(cap: &str) -> Result<Self, Self::Err> {
+        let cap = cap.replace('_', "").replace(',', ".");
+        match Percent::from_str(&cap) {
+            Ok(p) => {
+                if p.is_zero() {
+                    warn!("Pledge cap set to 0%, are you sure this is right?")
+                }
+                Ok(PledgeCap::Percent(p))
+            }
+            Err(_) => {
+                match cap.parse::<u128>() {
+                    Ok(i) => {
+                        if i < 100_000_000_000 {
+                            warn!("PledgeCap set to less then 100_000 NYM, are you sure this is right?");
+                        }
+                        Ok(PledgeCap::Absolute(Uint128::from(i)))
+                    }
+                    Err(_e) => Err(format!("Could not parse {} as Percent or Uint128", cap)),
+                }
+            }
+        }
+    }
+}
+
+impl Default for PledgeCap {
+    fn default() -> Self {
+        #[allow(clippy::expect_used)]
+        PledgeCap::Percent(Percent::from_percentage_value(10).expect("This can never fail!"))
     }
 }
 
@@ -101,4 +145,33 @@ pub struct DelegationTimesResponse {
 pub struct AllDelegationsResponse {
     pub delegations: Vec<VestingDelegation>,
     pub start_next_after: Option<(u32, MixId, u64)>,
+}
+
+#[cfg(test)]
+mod test {
+    use contracts_common::Percent;
+    use cosmwasm_std::Uint128;
+    use std::str::FromStr;
+
+    use crate::PledgeCap;
+
+    #[test]
+    fn test_pledge_cap_from_str() {
+        assert_eq!(
+            PledgeCap::from_str("0.1").unwrap(),
+            PledgeCap::Percent(Percent::from_percentage_value(10).unwrap())
+        );
+        assert_eq!(
+            PledgeCap::from_str("0,1").unwrap(),
+            PledgeCap::Percent(Percent::from_percentage_value(10).unwrap())
+        );
+        assert_eq!(
+            PledgeCap::from_str("100_000_000_000").unwrap(),
+            PledgeCap::Absolute(Uint128::new(100_000_000_000))
+        );
+        assert_eq!(
+            PledgeCap::from_str("100000000000").unwrap(),
+            PledgeCap::Absolute(Uint128::new(100_000_000_000))
+        );
+    }
 }