+++ common/nymsphinx/params/src/lib.rs
@@ -13,6 +13,7 @@ pub use packet_sizes::PacketSize;
 
 pub mod packet_modes;
 pub mod packet_sizes;
+pub mod packet_version;
 
 // If somebody can provide an argument why it might be reasonable to have more than 255 mix hops,
 // I will change this to [`usize`]
@@ -23,6 +24,17 @@ pub const DEFAULT_NUM_MIX_HOPS: u8 = 3;
 // access to that
 pub const FRAG_ID_LEN: usize = 5;
 pub type SerializedFragmentIdentifier = [u8; FRAG_ID_LEN];
+
+// wait, wait, but why are we starting with version 7?
+// when packet header gets serialized, the following bytes (in that order) are put onto the wire:
+// - packet_version (starting with v1.1.0)
+// - packet_size indicator
+// - packet_mode
+// it also just so happens that the only valid values for packet_size indicator include values 1-6
+// therefore if we receive byte `7` (or larger than that) we'll know we received a versioned packet,
+// otherwise we should treat it as legacy
+/// Increment it whenever we perform any breaking change in the wire format!
+const CURRENT_PACKET_VERSION_NUMBER: u8 = 7;
 
 // TODO: ask @AP about the choice of below algorithms