+++ explorer-api/src/mix_node/econ_stats.rs
@@ -4,7 +4,7 @@
 use crate::client::ThreadsafeValidatorClient;
 use crate::helpers::best_effort_small_dec_to_f64;
 use crate::mix_node::models::EconomicDynamicsStats;
-use mixnet_contract_common::rewarding::helpers::truncate_decimal;
+use contracts_common::truncate_decimal;
 use mixnet_contract_common::MixId;
 
 pub(crate) async fn retrieve_mixnode_econ_stats(