+++ nym-wallet/src/context/mocks/bonding.tsx
@@ -26,6 +26,7 @@ const bondedMixnodeMock: TBondedMixnode = {
   mixPort: 1789,
   verlocPort: 1790,
   version: '1.0.2',
+  isUnbonding: false,
 };
 
 const bondedGatewayMock: TBondedGateway = {
@@ -38,6 +39,11 @@ const bondedGatewayMock: TBondedGateway = {
   mixPort: 1789,
   verlocPort: 1790,
   version: '1.0.2',
+  routingScore: {
+    average: 100,
+    current: 100,
+  },
+  isUnbonding: false,
 };
 
 const TxResultMock: TransactionExecuteResult = {