+++ service-providers/network-requester/src/main.rs
@@ -5,6 +5,7 @@ use clap::CommandFactory;
 use clap::Subcommand;
 use clap::{Args, Parser};
 use completions::{fig_generate, ArgShell};
+use logging::setup_logging;
 
 use error::NetworkRequesterError;
 use network_defaults::DEFAULT_WEBSOCKET_LISTENING_PORT;
@@ -105,22 +106,4 @@ async fn main() -> Result<(), NetworkRequesterError> {
     let args = Cli::parse();
 
     execute(args).await
-}
-
-fn setup_logging() {
-    let mut log_builder = pretty_env_logger::formatted_timed_builder();
-    if let Ok(s) = ::std::env::var("RUST_LOG") {
-        log_builder.parse_filters(&s);
-    } else {
-        // default to 'Info'
-        log_builder.filter(None, log::LevelFilter::Info);
-    }
-
-    log_builder
-        .filter_module("hyper", log::LevelFilter::Warn)
-        .filter_module("tokio_reactor", log::LevelFilter::Warn)
-        .filter_module("reqwest", log::LevelFilter::Warn)
-        .filter_module("mio", log::LevelFilter::Warn)
-        .filter_module("want", log::LevelFilter::Warn)
-        .init();
 }